+++ tests/runtime/utils.py
@@ -3,10 +3,17 @@ import sys
 
 class Utils():
     @staticmethod
+    def assertRegexp(cls, a, b):
+        if sys.version_info[0] >= 3:
+            return cls.assertRegex(a, b)
+        else:
+            return cls.assertRegexpMatches(a, b)
+
+    @staticmethod
     def run_test(cls, cmd, expected):
         p = subprocess.Popen(
             [cmd], shell=True, stdout=subprocess.PIPE, stderr=subprocess.STDOUT)
-        cls.assertRegexpMatches(p.stdout.readline(), expected)
+        Utils.assertRegexp(cls, p.stdout.readline().decode('utf-8', 'ignore'), expected)
 
     @staticmethod
     def regex_uint(test_name):
